bot_service/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
langchain
langchain-community
langchain-deepseek
openai
python-dotenv